audio effects: add auxiliary channel feature
Added API to query and configure effects supporting auxiliary channels.
An auxiliary channel contains an additional audio signal helping in the
processing of the main channel. An example is a dual mic noise suppressor
were the main channel contains the signal from the main mic and the aux channel
containss the signal from the sub mic.

Change-Id: Ibb7a19d7f492720ebad0be45f78f996e0beba272