hardware: rename version_major/minor to module/hal api version
The previous names and documentation were not clear as to how the
version fields should be used. As a result, they were often either
unused or used improperly.

It became clear that there were two version fields necessary. One
to version the implementing module and the other to version the
hw_module_t interface itself.

This change renames version_major and version_minor members of
hw_module_t to module_api_version and hal_api_version. It also
provides in-depth descriptions for all the fields and clarifies
the purpose of the version field in hw_device_t structure.

Change-Id: I3e33e5a922cdc17c5e3b1c30a00e211394f18e86
Signed-off-by: Dima Zavin <dima@android.com>