hardware: provide temp version defines for source compatibility
This allows all the existing modules to have source level
compatibility until they are fixed to use the new field names.

Since the binary compatibility has not been altered, we do not
need to force all the implementors to rev their version numbers.

This change should be reverted when all the clients have been updated.

Change-Id: Ie543d3a972d8abe03ca5123e3000966f048d1771
Signed-off-by: Dima Zavin <dima@android.com>