Camera tests: Add variable burst test
By default, this test runs through a range of application-set
exposure, frame duration, and sensitivity values, and verifies
that the duration of capture, and in some cases the brightness,
match what's expected.

Optionally, it can use environment variables to specify the
burst parameters, and to enable dumping the captured YUV images
to flash for later debugging.

Change-Id: I647d360344ef3684e99c86c369e96ac82e62fc96