Update HDMI-CEC HAL interface definition
Previously it was HAL that was responsible for the logical address
allocation. The address allocation logic itself is based on the algorithm
specified in the CEC standard, is dependent neither on vendor nor
on hardware. This CL replaces the method 'allocate_logical_address',
'get_logical_address' with 'add_logical_address', so that the common
allocation logic will be implemented in the cec service and can be used by
all vendors without having them put the same logic in all HAL
implementations. HAL is supposed to be passed the allocated address
and use it to configure the chipset in this new definition.

The change in the service that implements the address allocation logic
will be done in a follow up CL.

Also removed the description in 'get_version' about version number.
It needs not be in the description since it is an implementation detail.

Change-Id: I33f31580792785bec38b7bd6997f91b89bfbb500