hwcomposer: Add function pointer for setPowerMode() callback
The blank HWC call is being replaced by a setPowerMode() call.
This incorporates all of the blank functionality, but provides added
support for display low power modes, if they are available in hardware.

Currently there are three supported modes:
- HWC_POWER_MODE_OFF    : Turn screen off.
- HWC_POWER_MODE_DOZE   : Put screen in low power mode.
- HWC_POWER_MODE_NORMAL : Turn screen on, and exit low power mode (if
                          display was previously in low power mode).

Where applicable, the device driver is expected to retain and restore
the low power mode state of the display during entry and exit from
suspend, respectively.

Bug: 13472578
Change-Id: I743ea3a2cb58247d14d7c47158b09a5309cca710
Signed-off-by: Prashant Malani <pmalani@google.com>