usb audio: Fix incorrect bytes returned for recording
In a recording session , the number of bytes read is
not updated when pcm_read() fails. This results in
silence/invalid data being added during a device switch
from USB headset to handset.
Fix is to reset the number of bytes read if pcm_read()
returns an error indicating removal of USB headset.

Bug: 18611518

Change-Id: I10d578c3cf1037c90a891e63be2bd1c2b7e6126b