Creating audio.stub.default for default HAL audio output
Creating a simple "stub" audio HAL to be used in systems with no
hardware output.  This stub HAL allows systems to run CTS tests and
other audio system calls without crashing or unexpected behavior.
The stub can be used also as a starting point for creating a new HAL.

This is needed for audio_output feature support.

bug: 19439530
Change-Id: I9e8e06e5b7fea7c7f3ff1de303db015b51e8acde