Add input HAL methods for getting device properties.
A property map holds the collection of the properties for a given
device identifier. HAL modules allocate and release these property maps
to control the lifetime of the data. The host will typically open and
read the corresponding idc file on allocation, and release these
resources when it is freed.

Change-Id: Ia13ec7b7abf1a79eeed87dc744c174a37f1fb8a9