context_hub: Fix version defines
We're defining the version of a hardware module (hw_module_t), not
a hardware device (hw_device_t).  Thus, we only have 16 bits for
our version number, instead of 32, and need to use the
HARDWARE_DEVICE_API_VERSION macro.

Additionally, for the CONTEXT_HUB_DEVICE_API_VERSION_1_0 define,
which has 1 and 0 hard-coded in the name itself, we hard code 1 and 0
within its definition.

We create CONTEXT_HUB_DEVICE_API_VERSION which will change as we update
the major and minor version numbers in this header.

Bug: 27798362
Change-Id: I58b9751dfb1989fd6bb7fbfc01329494cd88130c