Abstract V4L2 controls into wrapper class.
This abstraction will both clean up the code in V4L2Camera.cpp, and make it
easier to implement tests in the future (by mocking over the class).

This CL only adds the class and implementation, it does not yet replace
the logic in V4L2Camera.cpp.

Also gates building the module, to prevent checkbuilds without the
proper v4l2 kernel headers from trying to build it.

BUG: 30140438
Change-Id: I4e5feee30fdf896b1f71f0df492ca313f5581e78