Make HW_AUTH_TOKEN_VERSION static
Having this object non-static in header file is ok in
"c-plus-plus" since constant values default to internal
linkage, which allows them to appear in header files.

This is not i case in "c" and we need to declare it also
as static to limit the contex.

Making it static works for both.

Bug: 32467815
Test: lunch 1; make
Change-Id: I4ed4ffe41add3b10d990db0fb156a85a146d0ba2
Signed-off-by: Aleksej Makarov <aleksej.makarov@sonymobile.com>